# backend/internal/ws/ws_handler.go
package ws

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/gorilla/websocket"
)

type Handler struct {
	Hub *Hub
}

func NewHandler(h *Hub) *Handler {
	return &Handler{
		Hub: h,
	}
}

type CreateRoomReq struct {
	ID   string `json:"id"`
	Name string `json:"name"`
}

// CreateRoom godoc
// @Summary Create a room
// @Description Create a new WebSocket room
// @Tags websocket
// @Accept json
// @Produce json
// @Param room body CreateRoomReq true "Room data"
// @Success 200 {object} CreateRoomReq
// @Failure 500 {object} map[string]string
// @Router /ws/create-rooms [post]
func (h *Handler) CreateRoom(c *gin.Context) {
	var req CreateRoomReq
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"message": "Error occured in binding json"})
		return
	}
	h.Hub.Rooms[req.ID] = &Rooms{
		ID:      req.ID,
		Name:    req.Name,
		Clients: make(map[string]*Client),
	}
	c.JSON(http.StatusOK, req)

}

var upgrader = websocket.Upgrader{
	ReadBufferSize:  1024,
	WriteBufferSize: 1024,
	CheckOrigin: func(r *http.Request) bool {
		return true
	},
}

func (h *Handler) JoinRoom(c *gin.Context) {
	conn, err := upgrader.Upgrade(c.Writer, c.Request, nil)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"message": "Error in connecting the websocket"})
		return
	}
	roomId := c.Param("room_id")
	clientId := c.Query("user_id")
	username := c.Query("username")

	cl := &Client{
		Conn:     conn,
		RoomID:   roomId,
		Username: username,
		Message:  make(chan *Message),
		ID:       clientId,
	}
	m := &Message{
		RoomID:   roomId,
		Username: username,
		Content:  "A new user has joined the room",
	}
	h.Hub.Register <- cl
	h.Hub.Broadcast <- m

	go cl.writeMessage()
	cl.readMessage(h.Hub)

}

type RoomRes struct {
	ID   string `json:"id"`
	Name string `json:"name"`
}

// GetRooms godoc
// @Summary Get all rooms
// @Description Retrieve list of all rooms
// @Tags websocket
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Router /ws/get-rooms [post]
func (h *Handler) GetRooms(c *gin.Context) {
	rooms := make([]RoomRes, 0)

	for _, r := range h.Hub.Rooms {
		rooms = append(rooms, RoomRes{
			ID:   r.ID,
			Name: r.Name,
		})
	}
	c.JSON(http.StatusOK, gin.H{"message": "Room Response fetch success", "data": rooms})
}

type ClientRes struct {
	ID       string `json:"id"`
	Username string `json:"username"`
}

// GetClients godoc
// @Summary Get clients in a room
// @Description Retrieve list of clients in a specific room
// @Tags websocket
// @Produce json
// @Param roomId path string true "Room ID"
// @Success 200 {array} ClientRes
// @Router /ws/getClients/{roomId} [get]
func (h *Handler) GetClients(c *gin.Context) {
	var clients []ClientRes
	roomId := c.Param("roomId")

	if _, ok := h.Hub.Rooms[roomId]; !ok {
		clients = make([]ClientRes, 0)
		c.JSON(http.StatusOK, clients)
	}

	for _, c := range h.Hub.Rooms[roomId].Clients {
		clients = append(clients, ClientRes{
			ID:       c.ID,
			Username: c.Username,
		})
	}

	c.JSON(http.StatusOK, clients)
}